backend/requirements.txt
# Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Celery
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.1

# URL Validation
validators==0.22.0

# QR Code Generation
qrcode[pil]==7.4.2

# WSGI Server
gunicorn==21.2.0

# API Documentation
drf-spectacular==0.26.5

# Caching
django-redis==5.4.0

# Rate Limiting
django-ratelimit==4.1.0

# Utilities
python-dotenv==1.0.0
hashids==1.3.1
user-agents==2.2.0
